add swal for delete of goal

+++ onTrack.API/build/js/appAll.js
@@ -314,6 +314,40 @@
 }());
 (function() {
 	angular.module('onTrack')
+	.controller('MainController', ['$scope', '$state', '$http', '$window', 
+		function($scope, $state, $http, $window) {
+
+			function getTeams() {
+				$http.get('api/teams')
+					.success(function(data) {
+						$scope.teams = data
+					})
+					.error(function(err) {
+						console.log(err);
+					})
+			}
+
+			function getUsers() {
+				$http.get('api/users')
+					.success(function(data) {
+						$scope.users = data
+					})
+					.error(function(err) {
+						console.log(err)
+					})
+			}
+
+			$scope.logout = function() {
+				$window.sessionStorage.clear()
+				$state.go('login')
+			}
+		
+			getUsers() 
+			getTeams()
+	}])
+}());
+(function() {
+	angular.module('onTrack')
 	.controller('MainSettingsController', ['$scope', '$state', '$window', '$http',
 	 function($scope, $state, $window, $http) {
 
@@ -379,13 +413,136 @@
 }());
 (function() {
 	angular.module('onTrack')
-	.controller('MainController', ['$scope', '$state', '$http', '$window', 
-		function($scope, $state, $http, $window) {
+	.controller('TeamController', ['$scope', '$state', '$http', '$window', '$stateParams', '$q',
+		function($scope, $state, $http, $window, $stateParams, $q) {
+		
+			$scope.removeAdmin = function(admin) {
+				var token = $window.sessionStorage['jwt']
+
+				$http.put('/api/teams/' + $stateParams['id'] + '/removeAdmin', admin, {
+					headers: {
+						'Authorization': `Bearer ${token}`
+					}
+				})
+				.success(function(data){
+					$scope.team = data
+				})
+				.error(function(err) {
+					console.log(err)
+				})
+			}
+
+			$scope.removeUser = function(user) {
+				//this automatically makes two requests. Check to see if you need to. If not then only issue one update
+				var token = $window.sessionStorage['jwt']
+
+				$http.put('/api/teams/' + $stateParams.id + '/removeAdmin', user , {
+						headers: {
+							'Authorization': `Bearer ${token}`
+						}
+					})
+					.success(function(data) {
+						$http.put('/api/teams/' + $stateParams.id + '/removeUser', user, {
+								headers: {
+									'Authorization': `Bearer ${token}`
+								}
+							})
+							.success(function(data) {
+								$scope.team = data
+							})
+							.error(function(err) {
+								console.log(err)
+						})
+					})
+					.error(function(err) {
+						console.log(err)
+				})
+			}
+
+			$scope.deleteTeam = function() {
+				var token = $window.sessionStorage['jwt']
+
+				swal({
+				  title: "Are you sure?",
+				  text: "You will not be able to recover this team!",
+				  type: "warning",
+				  showCancelButton: true,
+				  confirmButtonColor: "#DD6B55",
+				  confirmButtonText: "Yes, delete it!",
+				  html: false
+				}).then(function(){
+					$http.delete('/api/teams/' + $stateParams['id'],
+					 {
+						headers: {
+							'Authorization': `Bearer ${token}`
+						}
+					})
+					.success(function(data){
+						$state.go('main')
+					})
+					.error(function(err) {
+						console.log(err)
+					})
+				}).catch(function(){
+					return
+				})
+			}
+
+			$scope.logout = function() {
+				$window.sessionStorage.clear()
+				$state.go('login')
+			}
+
+			$scope.updateName = function(data) {
+				if (data === ''){
+					return 'Name is required'
+				}
+			
+				if($scope.nameArr.includes(data)){
+					return 'Name is already taken'
+				}
+				return updateTeam(data, 'name')
+			}
+
+			function updateTeam(data, field) {
+				var d = $q.defer();
+				var token = $window.sessionStorage['jwt']
+				$scope.team[field] = data
+
+				$http.put('/api/teams/' + $stateParams['id'], $scope.team,{
+					headers: {
+						'Authorization': `Bearer ${token}`
+					}
+				})
+				.success(function(data){
+					$state.reload()
+				})
+				.error(function(err) {
+					console.log(err)
+				})
+
+			}
+
+
+
+			function getTeam(){
+				$http.get('/api/teams/' + $stateParams['id'])
+					.success(function(data) {
+						$scope.team = data
+					})
+					.error(function(err) {
+						console.log(err);
+					})
+			}
 
 			function getTeams() {
 				$http.get('api/teams')
 					.success(function(data) {
-						$scope.teams = data
+						$scope.nameArr = []
+						data.forEach(function(team){
+							$scope.nameArr.push(team.name)
+						})
+
 					})
 					.error(function(err) {
 						console.log(err);
@@ -395,20 +552,21 @@
 			function getUsers() {
 				$http.get('api/users')
 					.success(function(data) {
-						$scope.users = data
+						$scope.users = []
+						data.forEach(function(){
+							$scope.users.push({name: data.firstName + ' ' + data.lastName})
+						})
+
 					})
 					.error(function(err) {
 						console.log(err)
 					})
 			}
 
-			$scope.logout = function() {
-				$window.sessionStorage.clear()
-				$state.go('login')
-			}
-		
-			getUsers() 
 			getTeams()
+			getTeam()
+
+		
 	}])
 }());
 (function() {
@@ -683,164 +841,6 @@
 
 			populateTypes()
 			getUser()
-		
-	}])
-}());
-(function() {
-	angular.module('onTrack')
-	.controller('TeamController', ['$scope', '$state', '$http', '$window', '$stateParams', '$q',
-		function($scope, $state, $http, $window, $stateParams, $q) {
-		
-			$scope.removeAdmin = function(admin) {
-				var token = $window.sessionStorage['jwt']
-
-				$http.put('/api/teams/' + $stateParams['id'] + '/removeAdmin', admin, {
-					headers: {
-						'Authorization': `Bearer ${token}`
-					}
-				})
-				.success(function(data){
-					$scope.team = data
-				})
-				.error(function(err) {
-					console.log(err)
-				})
-			}
-
-			$scope.removeUser = function(user) {
-				//this automatically makes two requests. Check to see if you need to. If not then only issue one update
-				var token = $window.sessionStorage['jwt']
-
-				$http.put('/api/teams/' + $stateParams.id + '/removeAdmin', user , {
-						headers: {
-							'Authorization': `Bearer ${token}`
-						}
-					})
-					.success(function(data) {
-						$http.put('/api/teams/' + $stateParams.id + '/removeUser', user, {
-								headers: {
-									'Authorization': `Bearer ${token}`
-								}
-							})
-							.success(function(data) {
-								$scope.team = data
-							})
-							.error(function(err) {
-								console.log(err)
-						})
-					})
-					.error(function(err) {
-						console.log(err)
-				})
-			}
-
-			$scope.deleteTeam = function() {
-				var token = $window.sessionStorage['jwt']
-
-				swal({
-				  title: "Are you sure?",
-				  text: "You will not be able to recover this team!",
-				  type: "warning",
-				  showCancelButton: true,
-				  confirmButtonColor: "#DD6B55",
-				  confirmButtonText: "Yes, delete it!",
-				  html: false
-				}).then(function(){
-					$http.delete('/api/teams/' + $stateParams['id'],
-					 {
-						headers: {
-							'Authorization': `Bearer ${token}`
-						}
-					})
-					.success(function(data){
-						$state.go('main')
-					})
-					.error(function(err) {
-						console.log(err)
-					})
-				}).catch(function(){
-					return
-				})
-			}
-
-			$scope.logout = function() {
-				$window.sessionStorage.clear()
-				$state.go('login')
-			}
-
-			$scope.updateName = function(data) {
-				if (data === ''){
-					return 'Name is required'
-				}
-			
-				if($scope.nameArr.includes(data)){
-					return 'Name is already taken'
-				}
-				return updateTeam(data, 'name')
-			}
-
-			function updateTeam(data, field) {
-				var d = $q.defer();
-				var token = $window.sessionStorage['jwt']
-				$scope.team[field] = data
-
-				$http.put('/api/teams/' + $stateParams['id'], $scope.team,{
-					headers: {
-						'Authorization': `Bearer ${token}`
-					}
-				})
-				.success(function(data){
-					$state.reload()
-				})
-				.error(function(err) {
-					console.log(err)
-				})
-
-			}
-
-
-
-			function getTeam(){
-				$http.get('/api/teams/' + $stateParams['id'])
-					.success(function(data) {
-						$scope.team = data
-					})
-					.error(function(err) {
-						console.log(err);
-					})
-			}
-
-			function getTeams() {
-				$http.get('api/teams')
-					.success(function(data) {
-						$scope.nameArr = []
-						data.forEach(function(team){
-							$scope.nameArr.push(team.name)
-						})
-
-					})
-					.error(function(err) {
-						console.log(err);
-					})
-			}
-
-			function getUsers() {
-				$http.get('api/users')
-					.success(function(data) {
-						$scope.users = []
-						data.forEach(function(){
-							$scope.users.push({name: data.firstName + ' ' + data.lastName})
-						})
-
-					})
-					.error(function(err) {
-						console.log(err)
-					})
-			}
-
-			getTeams()
-			getTeam()
-
 		
 	}])
 }());
@@ -1237,82 +1237,6 @@
 
 	 		getSetting()
 
-	}])
-}());
-(function() {
-	angular.module('onTrack')
-	.controller('ProgressController', ['$scope', '$state', '$http', '$window', '$stateParams', '$q',
-		function($scope, $state, $http, $window, $stateParams, $q) {
-			
-			function getProgress() {
-				$http.get('api/users/' + $stateParams.id + '/progress/' + $stateParams.progressId)
-					.success(function(data) {
-						$scope.progress = data
-						$http.get('api/progress-settings/' + data.settingId)
-							.success(function(data){
-								$scope.setting = data
-							})
-							.error(function(err) {
-								console.log(err)
-							})
-					})
-					.error(function(err) {
-						console.log(err)
-					})
-			}
-
-
-			$scope.updateValue = function(data) {
-				if (data < 0) {
-					return "Value can't be negative"
-				}
-				updateProgress(data)
-			}
-
-			function updateProgress(data) {
-				$scope.progress['value'] = data
-
-				$http.put('api/users/' + $stateParams.id + '/progress/' + $stateParams.progressId, $scope.progress)
-					.error(function(err) {
-						console.log(err)
-					})
-
-			}
-
-			getProgress()
-
-
-	}])
-}());
-(function() {
-	angular.module('onTrack')
-	.controller('UserFormController', ['$scope', '$state', '$http', '$window', 
-		function($scope, $state, $http, $window) {
-
-			$scope.errorDisplay = false
-
-			$scope.createUser = function(user) {
-				
-				var duplicate = $scope.users.filter(function(x){
-					return x.email == user.email
-				})
-
-				if (duplicate.length > 0){
-					$scope.oops = 'Email is already taken!'
-					$scope.errorDisplay = true
-					return
-				}
-				else{
-					$http.post('api/users', user)
-						.success(function(data){
-							$state.reload()
-
-						})
-						.error(function(err) {
-							console.log(err)
-						})
-				}
-			}
 	}])
 }());
 (function() {
@@ -1839,6 +1763,82 @@
 			getPromotions()		
 	}])
 }());
+(function() {
+	angular.module('onTrack')
+	.controller('UserFormController', ['$scope', '$state', '$http', '$window', 
+		function($scope, $state, $http, $window) {
+
+			$scope.errorDisplay = false
+
+			$scope.createUser = function(user) {
+				
+				var duplicate = $scope.users.filter(function(x){
+					return x.email == user.email
+				})
+
+				if (duplicate.length > 0){
+					$scope.oops = 'Email is already taken!'
+					$scope.errorDisplay = true
+					return
+				}
+				else{
+					$http.post('api/users', user)
+						.success(function(data){
+							$state.reload()
+
+						})
+						.error(function(err) {
+							console.log(err)
+						})
+				}
+			}
+	}])
+}());
+(function() {
+	angular.module('onTrack')
+	.controller('ProgressController', ['$scope', '$state', '$http', '$window', '$stateParams', '$q',
+		function($scope, $state, $http, $window, $stateParams, $q) {
+			
+			function getProgress() {
+				$http.get('api/users/' + $stateParams.id + '/progress/' + $stateParams.progressId)
+					.success(function(data) {
+						$scope.progress = data
+						$http.get('api/progress-settings/' + data.settingId)
+							.success(function(data){
+								$scope.setting = data
+							})
+							.error(function(err) {
+								console.log(err)
+							})
+					})
+					.error(function(err) {
+						console.log(err)
+					})
+			}
+
+
+			$scope.updateValue = function(data) {
+				if (data < 0) {
+					return "Value can't be negative"
+				}
+				updateProgress(data)
+			}
+
+			function updateProgress(data) {
+				$scope.progress['value'] = data
+
+				$http.put('api/users/' + $stateParams.id + '/progress/' + $stateParams.progressId, $scope.progress)
+					.error(function(err) {
+						console.log(err)
+					})
+
+			}
+
+			getProgress()
+
+
+	}])
+}());
 var app = angular.module('app', ['autocomplete']);
 
 // the service that retrieves some movie title from an url
@@ -2298,13 +2298,24 @@ app.directive('suggestion', function(){
 	function($scope, $state, $http, $stateParams) {
 
 		$scope.deleteGoal = function() {
-			$http.delete('api/goals/' + $stateParams.id)
+
+			swal({
+				title: "Are you sure?",
+				text: "You will not be able to recover this goal",
+				type: "warning",
+				showCancelButton: true,
+				confirmButtonColor: "#DD6B55",
+				confirmButtonText: "Yes, delete it!",
+				html: false
+			}).then(function() {
+				$http.delete('api/goals/' + $stateParams.id)
 				.success(function(data) {
 					$state.go('goals')
 				})
 				.error(function(err) {
 					console.log(err)
 				})
+			}).catch(e => {})
 		}
 	}])
